Add structcheck, unused, and varcheck linters.
Warn on unused and dead code
