+++ .gometalinter.json
@@ -10,6 +10,10 @@
   "WarnUnmatchedDirective": true,
 
   "Enable": [
+    "structcheck",
+    "unused",
+    "varcheck",
+
     "gofmt",
     "goimports",
     "golint",
