--- archive/path.go
@@ -1,1 +1,0 @@
-package archive
