+++ cmd/containerd-shim/main_unix.go
@@ -29,16 +29,6 @@ import (
 	"google.golang.org/grpc"
 )
 
-const usage = `
-                    __        _                     __           __    _
-  _________  ____  / /_____ _(_)___  ___  _________/ /     _____/ /_  (_)___ ___
- / ___/ __ \/ __ \/ __/ __ ` + "`" + `/ / __ \/ _ \/ ___/ __  /_____/ ___/ __ \/ / __ ` + "`" + `__ \
-/ /__/ /_/ / / / / /_/ /_/ / / / / /  __/ /  / /_/ /_____(__  ) / / / / / / / / /
-\___/\____/_/ /_/\__/\__,_/_/_/ /_/\___/_/   \__,_/     /____/_/ /_/_/_/ /_/ /_/
-
-shim for container lifecycle and reconnection
-`
-
 var (
 	debugFlag            bool
 	namespaceFlag        string
