+++ cmd/containerd/main.go
@@ -8,7 +8,6 @@ import (
 	"net"
 	"os"
 	"os/signal"
-	"runtime"
 	"time"
 
 	"google.golang.org/grpc/grpclog"
@@ -195,19 +194,4 @@ func setLevel(context *cli.Context, config *server.Config) error {
 		logrus.SetLevel(lvl)
 	}
 	return nil
-}
-
-func dumpStacks() {
-	var (
-		buf       []byte
-		stackSize int
-	)
-	bufferLen := 16384
-	for stackSize == len(buf) {
-		buf = make([]byte, bufferLen)
-		stackSize = runtime.Stack(buf, true)
-		bufferLen *= 2
-	}
-	buf = buf[:stackSize]
-	logrus.Infof("=== BEGIN goroutine stack dump ===\n%s\n=== END goroutine stack dump ===", buf)
 }
