+++ content/local/store_unix.go
@@ -10,15 +10,6 @@ import (
 	"github.com/containerd/containerd/sys"
 )
 
-func getStartTime(fi os.FileInfo) time.Time {
-	if st, ok := fi.Sys().(*syscall.Stat_t); ok {
-		return time.Unix(int64(sys.StatCtime(st).Sec),
-			int64(sys.StatCtime(st).Nsec))
-	}
-
-	return fi.ModTime()
-}
-
 func getATime(fi os.FileInfo) time.Time {
 	if st, ok := fi.Sys().(*syscall.Stat_t); ok {
 		return time.Unix(int64(sys.StatAtime(st).Sec),
