+++ content/local/store_windows.go
@@ -5,10 +5,6 @@ import (
 	"time"
 )
 
-func getStartTime(fi os.FileInfo) time.Time {
-	return fi.ModTime()
-}
-
 func getATime(fi os.FileInfo) time.Time {
 	return fi.ModTime()
 }
