+++ gc/gc_test.go
@@ -108,47 +108,3 @@ func toNodes(s []string) []Node {
 	}
 	return n
 }
-
-func newScanner(refs []string) *stringScanner {
-	return &stringScanner{
-		i: -1,
-		s: refs,
-	}
-}
-
-type stringScanner struct {
-	i int
-	s []string
-}
-
-func (ss *stringScanner) Next() bool {
-	ss.i++
-	return ss.i < len(ss.s)
-}
-
-func (ss *stringScanner) Node() Node {
-	return Node{
-		Key: ss.s[ss.i],
-	}
-}
-
-func (ss *stringScanner) Cleanup() error {
-	ss.s[ss.i] = ""
-	return nil
-}
-
-func (ss *stringScanner) Err() error {
-	return nil
-}
-
-func (ss *stringScanner) All() []Node {
-	remaining := make([]Node, 0, len(ss.s))
-	for _, s := range ss.s {
-		if s != "" {
-			remaining = append(remaining, Node{
-				Key: s,
-			})
-		}
-	}
-	return remaining
-}
