+++ helpers_windows_test.go
@@ -52,19 +52,3 @@ func withNewSnapshot(id string, i Image) NewContainerOpts {
 		return nil
 	}
 }
-
-func withRemappedSnapshot(id string, i Image, u, g uint32) NewContainerOpts {
-	return func(ctx context.Context, client *Client, c *containers.Container) error {
-		return nil
-	}
-}
-
-func withRemappedSnapshotView(id string, i Image, u, g uint32) NewContainerOpts {
-	return func(ctx context.Context, client *Client, c *containers.Container) error {
-		return nil
-	}
-}
-
-func withNoop(_ context.Context, _ *Client, _ *containers.Container, _ *specs.Spec) error {
-	return nil
-}
