+++ progress/escape.go
@@ -3,6 +3,6 @@ package progress
 const (
 	escape = "\x1b"
 	reset  = escape + "[0m"
-	red    = escape + "[31m"
+	red    = escape + "[31m" // nolint: unused, varcheck
 	green  = escape + "[32m"
 )
